Add new modules for display, encrypted block, peripheral safety, vault header store, and virtual MSC
- Introduced `display.h` and `display.c` for managing UI display operations.
- Added `encrypted_block.h` and `encrypted_block.c` to handle encrypted block storage with integrity checks.
- Created `peripheral_safety.h` and `peripheral_safety.c` to ensure safe operation of USB connectors and removable blocks.
- Implemented `vault_header_store.h` and `vault_header_store.c` for serialization and deserialization of vault headers.
- Developed `virtual_msc.h` and `virtual_msc.c` to provide a virtual mass storage controller interface.
- Implemented block slice functionality in `block_slice.c` for managing slices of block devices.
- Enhanced error handling and state management across all new modules.

+++ firmware/host/otp_file.c
@@ -84,8 +84,8 @@ static bool read_rows(fv_device_roots_storage_t *storage, uint16_t first_row,
         const off_t offset = (off_t)(((size_t)first_row + index) * 2u);
         ok = read_all(descriptor, encoded, sizeof(encoded), offset);
         if (ok) {
-            output[index] = (uint16_t)encoded[0] |
-                            ((uint16_t)encoded[1] << 8u);
+            output[index] = (uint16_t)((uint16_t)encoded[0] |
+                            (uint16_t)((uint16_t)encoded[1] << 8u));
         }
     }
     if (close(descriptor) != 0) ok = false;
@@ -103,8 +103,8 @@ static bool write_rows(fv_device_roots_storage_t *storage, uint16_t first_row,
         uint8_t encoded[2] = {0};
         const off_t offset = (off_t)(((size_t)first_row + index) * 2u);
         ok = read_all(descriptor, encoded, sizeof(encoded), offset);
-        const uint16_t existing = (uint16_t)encoded[0] |
-                                  ((uint16_t)encoded[1] << 8u);
+        const uint16_t existing = (uint16_t)((uint16_t)encoded[0] |
+                                  (uint16_t)((uint16_t)encoded[1] << 8u));
         if (ok && (existing & input[index]) != existing) ok = false;
         const uint16_t programmed = existing | input[index];
         encoded[0] = (uint8_t)programmed;
